requirements.txt
# Core dependencies
pyside6==6.5.0
numpy==1.24.0
pandas==2.0.0
pyyaml==6.0.0
deeplabcut==2.3.0

# Data processing
scipy>=1.10.0
h5py>=3.8.0

# Image processing
opencv-python>=4.7.0
pillow>=9.5.0

# Development
pytest>=7.3.1
pytest-qt>=4.2.0
mypy>=1.3.0
black>=23.3.0

# Logging and config
structlog>=23.1.0 